% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{fpkm}
\alias{fpkm}
\title{An example of FPKM data}
\format{A data frame which contains the fpkm data. Each row is a gene and 
each column is a sample}
\usage{
fpkm
}
\description{
An example of FPKM data
}
\keyword{datasets}
