% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{rawcount}
\alias{rawcount}
\title{An example of RNASeq counts data}
\format{A data frame which contains the RNASeq counts data. Each row is a 
gene and each column is a sample}
\usage{
rawcount
}
\description{
An example of RNASeq counts data
}
\keyword{datasets}
